Move RPC non exclusive code out of magirpc.h
Note: The code has been moved to util.h not because of well thought design decision.
It has been moved to util.h so that I don't have to deal with Makefile while dealing with this.

Please check if util.h is actually the best place to put this code before merging this into production code.

And, well, remove this message on the rewritten code.